Fix lint issues and clean up
