+++ src/components/DifficultyMeter.jsx
@@ -1,23 +1,6 @@
 import React from 'react';
 import './DifficultyMeter.css';
-
-export const difficultyMap = {
-    'Beginner': { color: '#4DB6AC' },
-    'Basic': { color: '#FDD835' },
-    'Difficult': { color: '#F44336' },
-    'Expert': { color: '#8BC34A' },
-    'Challenge': { color: '#BA68C8' },
-};
-
-export const difficultyLevels = ['Beginner', 'Basic', 'Difficult', 'Expert', 'Challenge'];
-
-export const difficultyNameMapping = {
-    'Beginner': ['beginner'],
-    'Basic': ['basic', 'easy', 'light'],
-    'Difficult': ['difficult', 'medium', 'standard'],
-    'Expert': ['expert', 'hard', 'heavy'],
-    'Challenge': ['challenge', 'oni'],
-};
+import { difficultyMap } from '../utils/difficulties.js';
 
 export const DifficultyMeter = ({ level, difficultyName, isMissing, onClick, isSelected }) => {
     const style = {
@@ -33,3 +16,4 @@ export const DifficultyMeter = ({ level, difficultyName, isMissing, onClick, isS
         </div>
     );
 };
+
