+++ src/components/StepchartPage.jsx
@@ -1,12 +1,8 @@
-import React, { useEffect, useState, useMemo, useContext } from "react";
-import { useLocation } from 'react-router-dom';
+import React, { useEffect, useState, useMemo } from "react";
 
 import { ToggleBar } from "./ToggleBar";
 import { StepchartSection } from "./StepchartSection";
-import { DifficultyMeter, difficultyLevels, difficultyNameMapping } from './DifficultyMeter';
-import { SettingsContext } from '../contexts/SettingsContext.jsx';
-import { useFilters } from '../contexts/FilterContext.jsx';
-import { getBpmRange } from '../BPMTool.jsx';
+import { DifficultyMeter } from './DifficultyMeter';
 
 import styles from "./StepchartPage.module.css";
 import "../BPMTool.css";
@@ -18,8 +14,6 @@ const sectionSizesInMeasures = {
   2.5: 3,
   3: 3,
 };
-
-const HEADER_ID = "stepchart-page-header";
 
 // function scrollTargetBeatJustUnderHeader(beatId, headerId) {
 //   setTimeout(() => {
@@ -37,18 +31,10 @@ const HEADER_ID = "stepchart-page-header";
 export function StepchartPage({
   simfile,
   currentType: initialCurrentType,
-  setCurrentChart,
-  isCollapsed,
-  setIsCollapsed,
-  playStyle,
   speedmod,
 }) {
   const [currentType, setCurrentType] = useState(initialCurrentType);
-  const location = useLocation();
   const isLoading = !simfile;
-  const { targetBPM, multipliers } = useContext(SettingsContext);
-  const { filters } = useFilters();
-  const [showAltBpm, setShowAltBpm] = useState(false);
 
   useEffect(() => {
     setCurrentType(initialCurrentType);
@@ -57,7 +43,6 @@ export function StepchartPage({
   // useEffect(() => {
   //   const hash = (window.location.hash ?? "").replace("#", "");
   //   if (hash) {
-  //     scrollTargetBeatJustUnderHeader(hash, HEADER_ID);
   //   }
   // }, [location.hash]);
 
@@ -70,44 +55,15 @@ export function StepchartPage({
     charts: {}
   };
 
-  const calculation = useMemo(() => {
-    if (!targetBPM || !displaySimfile.displayBpm || displaySimfile.displayBpm === 'N/A') return null;
-    const numericTarget = Number(targetBPM) || 0;
-    const bpmRange = getBpmRange(displaySimfile.displayBpm);
-    if (bpmRange.max === 0) return null;
-    const idealMultiplier = numericTarget / bpmRange.max;
-    const closestMultiplier = multipliers.reduce((prev, curr) => Math.abs(curr - idealMultiplier) < Math.abs(prev - idealMultiplier) ? curr : prev);
-    const closestIndex = multipliers.indexOf(closestMultiplier);
-    const primarySpeed = (bpmRange.max * closestMultiplier);
-    let alternativeMultiplier = null;
-    if (primarySpeed > numericTarget) {
-        if (closestIndex > 0) alternativeMultiplier = multipliers[closestIndex - 1];
-    } else {
-        if (closestIndex < multipliers.length - 1) alternativeMultiplier = multipliers[closestIndex + 1];
-    }
-    const result = {
-        primary: { modifier: closestMultiplier, minSpeed: Math.round(bpmRange.min * closestMultiplier), maxSpeed: Math.round(primarySpeed), isRange: bpmRange.min !== bpmRange.max },
-        alternative: null
-    };
-    if (alternativeMultiplier) {
-        const altMaxSpeed = (bpmRange.max * alternativeMultiplier);
-        result.alternative = { modifier: alternativeMultiplier, minSpeed: Math.round(bpmRange.min * alternativeMultiplier), maxSpeed: Math.round(altMaxSpeed), isRange: bpmRange.min !== bpmRange.max, direction: altMaxSpeed > primarySpeed ? 'up' : 'down' };
-    }
-    if (result.alternative && result.primary.maxSpeed === result.alternative.maxSpeed) {
-        result.alternative = null;
-    }
-    return result;
-  }, [targetBPM, displaySimfile.displayBpm, multipliers]);
-
   const currentTypeMeta = displaySimfile.availableTypes.find(
     (at) => at.slug === currentType
   );
 
   const chart = currentTypeMeta ? displaySimfile.charts[currentType] : null;
-  
+
   const sectionGroups = useMemo(() => {
     if (!chart) return [];
-    
+
     const { arrows, freezes } = chart;
     const lastArrowOffset = (arrows[arrows.length - 1]?.offset ?? 0) + 0.25;
     const lastFreezeOffset = freezes[freezes.length - 1]?.endOffset ?? 0;
@@ -123,13 +79,12 @@ export function StepchartPage({
           startOffset={i}
           endOffset={Math.min(totalSongHeight, i + sectionSizesInMeasures[speedmod])}
           style={{ zIndex: Math.round(totalSongHeight) - i }}
-          headerId={HEADER_ID}
         />
       );
     }
 
     const groups = [];
-    const sectionsPerChunk = currentType.includes("single") ? 7 : 4;
+    const sectionsPerChunk = currentType.includes('single') ? 7 : 4;
     while (sections.length) {
       const sectionChunk = sections.splice(0, sectionsPerChunk);
       groups.push(
@@ -146,50 +101,10 @@ export function StepchartPage({
   }, [chart, speedmod, currentType]);
 
 
-  const title = currentTypeMeta ? `${
-    displaySimfile.title.translitTitleName || displaySimfile.title.titleName
-  } - ${currentType.replace(/-/g, ", ")} (${currentTypeMeta.feet})` : displaySimfile.title.titleName;
+  const title = currentTypeMeta
+    ? `${displaySimfile.title.translitTitleName || displaySimfile.title.titleName} - ${currentType.replace(/-/g, ", ")} (${currentTypeMeta.feet})`
+    : displaySimfile.title.titleName;
 
-  const renderDifficulties = (style) => {
-    const chartDifficulties = displaySimfile.availableTypes.filter(t => t.mode === style);
-
-    return difficultyLevels.map(levelName => {
-        const type = chartDifficulties.find(t => t.difficulty === levelName.toLowerCase());
-        let filteredOut = false;
-        if (type) {
-            if (filters.difficultyMin !== '' && type.feet < Number(filters.difficultyMin)) {
-                filteredOut = true;
-            }
-            if (filters.difficultyMax !== '' && type.feet > Number(filters.difficultyMax)) {
-                filteredOut = true;
-            }
-            if (filters.difficultyNames && filters.difficultyNames.length > 0) {
-                const lowerCaseFilterNames = filters.difficultyNames.map(n => n.toLowerCase());
-                if (!lowerCaseFilterNames.includes(type.difficulty.toLowerCase())) {
-                    filteredOut = true;
-                }
-            }
-        }
-
-        return (
-            <DifficultyMeter
-                key={`${style}-${levelName}`}
-                level={!type || filteredOut ? 'X' : type.feet}
-                difficultyName={levelName}
-                isMissing={!type || filteredOut}
-                isSelected={type && !filteredOut && type.slug === currentType}
-                onClick={() => {
-                  if (type && !filteredOut) {
-                    setCurrentType(type.slug);
-                    if (setCurrentChart) {
-                        setCurrentChart(type);
-                    }
-                  }
-                }}
-            />
-        );
-    });
-  };
 
   return (
     <>
