+++ src/utils/difficulties.js
@@ -1,0 +1,17 @@
+export const difficultyMap = {
+    Beginner: { color: '#4DB6AC' },
+    Basic: { color: '#FDD835' },
+    Difficult: { color: '#F44336' },
+    Expert: { color: '#8BC34A' },
+    Challenge: { color: '#BA68C8' },
+};
+
+export const difficultyLevels = ['Beginner', 'Basic', 'Difficult', 'Expert', 'Challenge'];
+
+export const difficultyNameMapping = {
+    Beginner: ['beginner'],
+    Basic: ['basic', 'easy', 'light'],
+    Difficult: ['difficult', 'medium', 'standard'],
+    Expert: ['expert', 'hard', 'heavy'],
+    Challenge: ['challenge', 'oni'],
+};
